requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
tortoise-orm[asyncpg]==0.20.0
pydantic==2.5.3
anyio==4.2.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
tzlocal==5.2